Add function addExercise in ExercisesViewModel to add Exercises to database.
Add parameter exerciseRepo in ExercisesViewModel to be able to call
ExerciseRepo for database operations.

Add extension function to CreationExtras in AppViewModelProvider to get
instance of WorkoutApplication. This will be used to get AppDataContainer
which in turn provides instances of all the repositories in the app.

Add exercises in ExercisesScreen to populate the list in ExercisesBody.